Update nsupdate_updater.sh
A temporary file has been added because the Conjob runs every 5 minutes and you get into the problem that the provider NSUPdate.info blocks the DNS update due to too frequent requests without IP change.

The Bash script creates the lastip.txt file itself.

+++ nsupdate_updater.sh
@@ -4,9 +4,13 @@
 interface=ppp0
 ipv4regex="\b([0-9]{1,3}\.){3}[0-9]{1,3}\b"
 ipv6regex="([0-9]|[a-f]|:){4,}"
+lastipfile=/etc/lastip.txt
 
-updateurlbaseipv4="https://DYNHOSTNAME.nsupdate.info:NSUPDATESECRET@ipv4.nsupdate.info/nic/update?myip="
-updateurlbaseipv6="https://DYNHOSTNAME.nsupdate.info:NSUPDATESECRET@ipv4.nsupdate.info/nic/update?myip="
+lastipv4=$(cat /etc/lastip.txt | egrep -i -o $ipv4regex)
+lastipv6=$(cat /etc/lastip.txt | egrep -i -o $ipv6regex)
+
+updateurlbaseipv4="https://NSUPDATEHOSTNAME.nsupdate.info:NSUPDATESECRET@ipv4.nsupdate.info/nic/update?myip="
+updateurlbaseipv6="https://NSUPDATEHOSTNAME.nsupdate.info:NSUPDATESECRET@ipv4.nsupdate.info/nic/update?myip="
 
 currentipv4=$(ip -f inet addr show $interface | egrep -i -o "inet.*peer" | egrep -i -o $ipv4regex)
 currentipv6=$(ip -f inet6 addr show $interface | egrep -i -o "inet6.*scope global dynamic" | egrep -i -o $ipv6regex)
@@ -14,5 +18,22 @@ currentipv6=$(ip -f inet6 addr show $interface | egrep -i -o "inet6.*scope globa
 updateurlipv4=$updateurlbaseipv4$currentipv4
 updateurlipv6=$updateurlbaseipv6$currentipv6
 
-wget -q -O --no-check-certificate "${updateurlipv4}"
-wget -q -O --no-check-certificate "${updateurlipv6}"
+if [ -f "$lastipfile" ]; then
+
+	if [ "$currentipv4" == "$lastipv4" ]; then
+		echo "IPv4 ist gleich"
+	else
+		sed -i "s/$lastipv4/$currentipv4/g" /etc/lastip.txt
+    wget -q -O --no-check-certificate "${updateurlipv4}"
+	fi
+
+	if [ "$currentipv6" == "$lastipv6" ]; then
+		echo "IPv6 ist gleich"
+	else
+		sed -i "s/$lastipv6/$currentipv6/g" /etc/lastip.txt
+    wget -q -O --no-check-certificate "${updateurlipv6}"
+	fi
+
+else
+	echo -e "ipv4=1.1.1.1\nipv6=2001::1" > $lastipfile
+fi
